#37287 #37353 (2.2.0.89) 型の置換を継続中。 / Ongoing type replacement.

+++ src/scores.c
@@ -622,8 +622,8 @@ errr predict_score(void)
 	sprintf(the_score.p_a, "%2d", p_ptr->pseikaku);
 
 	/* Save the level and such */
-	sprintf(the_score.cur_lev, "%3d", p_ptr->lev);
-	sprintf(the_score.cur_dun, "%3d", dun_level);
+	sprintf(the_score.cur_lev, "%3d", (int)p_ptr->lev);
+	sprintf(the_score.cur_dun, "%3d", (int)dun_level);
 	sprintf(the_score.max_lev, "%3d", p_ptr->max_plv);
 	sprintf(the_score.max_dun, "%3d", max_dlv[dungeon_type]);
